Rename KQL -> Kql to match acronym capitlization guideline.
